+++ Sift/SiftIosAppState.h
@@ -4,17 +4,18 @@
 @import CoreMotion;
 @import Foundation;
 
+#import "SiftHtDictionary.h"
 #import "SiftUtils.h"
 
-NSMutableDictionary *SFMakeEmptyIosAppState(void);
+SiftHtDictionary *SFMakeEmptyIosAppState(void);
 
-NSMutableDictionary *SFCollectIosAppState(CLLocationManager *locationManager, NSString *title);
+SiftHtDictionary *SFCollectIosAppState(CLLocationManager *locationManager, NSString *title) NS_EXTENSION_UNAVAILABLE_IOS("SFCollectIosAppState is not supported for iOS extensions.");
 
-NSDictionary *SFCLHeadingToDictionary(CLHeading *heading);
+SiftHtDictionary *SFCLHeadingToDictionary(CLHeading *heading);
 
-NSDictionary *SFCLLocationToDictionary(CLLocation *data);
+SiftHtDictionary *SFCLLocationToDictionary(CLLocation *data);
 
-NSDictionary *SFCMDeviceMotionToDictionary(CMDeviceMotion *data, SFTimestamp now);
-NSDictionary *SFCMAccelerometerDataToDictionary(CMAccelerometerData *data, SFTimestamp now);
-NSDictionary *SFCMGyroDataToDictionary(CMGyroData *data, SFTimestamp now);
-NSDictionary *SFCMMagnetometerDataToDictionary(CMMagnetometerData *data, SFTimestamp now);
+SiftHtDictionary *SFCMDeviceMotionToDictionary(CMDeviceMotion *data, SFTimestamp now);
+SiftHtDictionary *SFCMAccelerometerDataToDictionary(CMAccelerometerData *data, SFTimestamp now);
+SiftHtDictionary *SFCMGyroDataToDictionary(CMGyroData *data, SFTimestamp now);
+SiftHtDictionary *SFCMMagnetometerDataToDictionary(CMMagnetometerData *data, SFTimestamp now);