+++ Sift/SiftIosAppStateCollector.m
@@ -202,7 +202,7 @@ static const NSTimeInterval SF_MOTION_SENSOR_INTERVAL = 0.5;  // Unit: second.
                     [self stopMotionSensors];
                     
                     if ([self canCollectLocationData] && self->_locationManager.location) {
-                        [event.iosAppState setObject:SFCLLocationToDictionary(self->_locationManager.location) forKey:@"location"];
+                        [event.iosAppState setEntry:@"location" value:SFCLLocationToDictionary(self->_locationManager.location).entries];
                     }
                     
                     // Read heading before we stop location manager (it nullifies heading when stopped).
@@ -210,27 +210,27 @@ static const NSTimeInterval SF_MOTION_SENSOR_INTERVAL = 0.5;  // Unit: second.
                     [self->_locationManager stopUpdatingHeading];
                     
                     if (heading) {
-                        [event.iosAppState setObject:SFCLHeadingToDictionary(heading) forKey:@"heading"];
+                        [event.iosAppState setEntry:@"heading" value:SFCLHeadingToDictionary(heading).entries];
                     }
                     
                     [self addReadingsToIosAppState:event.iosAppState];
                     
-                    SF_DEBUG(@"iosAppState: %@", event.iosAppState);
+                    SF_DEBUG(@"iosAppState: %@", event.iosAppState.entries);
                     [Sift.sharedInstance appendEvent:event];
                 });
             } else {
                 if ([self canCollectLocationData] && self->_locationManager.location) {
-                    [event.iosAppState setObject:SFCLLocationToDictionary(self->_locationManager.location) forKey:@"location"];
+                    [event.iosAppState setEntry:@"location" value:SFCLLocationToDictionary(self->_locationManager.location).entries];
                 }
                 
                 CLHeading *heading = self->_locationManager.heading;
                 if (heading) {
-                    [event.iosAppState setObject:SFCLHeadingToDictionary(heading) forKey:@"heading"];
+                    [event.iosAppState setEntry:@"heading" value:SFCLHeadingToDictionary(heading).entries];
                 }
                 
                 [self addReadingsToIosAppState:event.iosAppState];
                 
-                SF_DEBUG(@"iosAppState: %@", event.iosAppState);
+                SF_DEBUG(@"iosAppState: %@", event.iosAppState.entries);
                 [Sift.sharedInstance appendEvent:event];
             }
             
@@ -439,22 +439,22 @@ static NSString * const SF_LAST_COLLECTED_AT = @"lastCollectedAt";
     }
 }
 
-- (void)addReadingsToIosAppState:(NSMutableDictionary *)iosAppState {
+- (void)addReadingsToIosAppState:(SiftHtDictionary *)iosAppState {
     SF_GENERICS(NSArray, NSDictionary *) *motion = [self convertReadings:_deviceMotionReadings converter:SFCMDeviceMotionToDictionary];
     if (motion.count) {
-        [iosAppState setObject:motion forKey:@"motion"];
+        [iosAppState setEntry:@"motion" value:motion];
     }
     SF_GENERICS(NSArray, NSDictionary *) *rawAccelerometer = [self convertReadings:_accelerometerReadings converter:SFCMAccelerometerDataToDictionary];
     if (rawAccelerometer.count) {
-        [iosAppState setObject:rawAccelerometer forKey:@"raw_accelerometer"];
+        [iosAppState setEntry:@"raw_accelerometer" value:rawAccelerometer];
     }
     SF_GENERICS(NSArray, NSDictionary *) *rawGyro = [self convertReadings:_gyroReadings converter:SFCMGyroDataToDictionary];
     if (rawGyro.count) {
-        [iosAppState setObject:rawGyro forKey:@"raw_gyro"];
+        [iosAppState setEntry:@"raw_gyro" value:rawGyro];
     }
     SF_GENERICS(NSArray, NSDictionary *) *rawMagnetometer = [self convertReadings:_magnetometerReadings converter:SFCMMagnetometerDataToDictionary];
     if (rawMagnetometer.count) {
-        [iosAppState setObject:rawMagnetometer forKey:@"raw_magnetometer"];
+        [iosAppState setEntry:@"raw_magnetometer" value:rawMagnetometer];
     }
 }
 
@@ -465,7 +465,7 @@ static NSString * const SF_LAST_COLLECTED_AT = @"lastCollectedAt";
         for (CMLogItem *reading in [buffer shallowCopy]) {
             // CMLogItem records timestamp since device boot.
             SFTimestamp timestamp = [[uptime dateByAddingTimeInterval:reading.timestamp] timeIntervalSince1970] * 1000;
-            [readings addObject:((NSDictionary *(*)(CMLogItem *, SFTimestamp))converter)(reading, timestamp)];
+            [readings addObject:((SiftHtDictionary *(*)(CMLogItem *, SFTimestamp))converter)(reading, timestamp).entries];
         }
         [buffer removeAllObjects];
     }