+++ Sift/SiftIosDeviceProperties.h
@@ -2,6 +2,8 @@
 
 @import Foundation;
 
-NSMutableDictionary *SFMakeEmptyIosDeviceProperties(void);
+#import "SiftHtDictionary.h"
 
-NSMutableDictionary *SFCollectIosDeviceProperties(void);
+SiftHtDictionary *SFMakeEmptyIosDeviceProperties(void);
+
+SiftHtDictionary *SFCollectIosDeviceProperties(void) NS_EXTENSION_UNAVAILABLE_IOS("SFCollectIosDeviceProperties is not supported for iOS extensions.");