+++ Sift/SiftIosDevicePropertiesCollector.h
@@ -3,8 +3,9 @@
 @import Foundation;
 
 /** Collect device properties. */
+NS_EXTENSION_UNAVAILABLE_IOS("SiftIosDevicePropertiesCollector is not supported for iOS extensions.")
 @interface SiftIosDevicePropertiesCollector : NSObject
 
-// Nothing here - a collector is autonomous.
+- (void)collect;
 
 @end