+++ Sift/SiftUploader.h
@@ -4,6 +4,7 @@
 
 #import "Sift.h"
 
+NS_EXTENSION_UNAVAILABLE_IOS("SiftUploader is not supported for iOS extensions.")
 @interface SiftUploader : NSObject <NSURLSessionTaskDelegate, NSURLSessionDataDelegate>
 
 - (instancetype)initWithArchivePath:(NSString *)archivePath sift:(Sift *)sift;