+++ Sources/sift-ios/sift_ios.swift
@@ -1,7 +1,7 @@
 import Foundation
 import Sift
 //import Sift.Swift.h
-
+@available(iOSApplicationExtension, unavailable)
 public class SiftManager {
     public static let sharedInstance: Sift = Sift.sharedInstance()
     public init() {